refactor(base): migrate base functionality to new shell structure
