+++ test/widgets/pi_hole_client_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:pi_hole_client/base.dart';
 import 'package:pi_hole_client/pi_hole_client.dart';
 import 'package:pi_hole_client/ui/home/home.dart';
+import 'package:pi_hole_client/ui/shell/base.dart';
 
 import './helpers.dart';
